WrapperView for Events implemented
