+++ src/Core/src/Platform/Gtk/BorderView.cs
@@ -23,7 +23,7 @@ public class BorderView : Frame, ICrossPlatformLayoutBacking
 			}
 			else if (value != null)
 			{
-				if (value.Parent is EventBoxWrapperView)
+				if (value.Parent is WrapperView)
 				{
 					Child = value.Parent;
 				}
